+++ app/Http/Controllers/FleetController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Fleet;
+use App\Models\Fleet;
 use Illuminate\Http\Request;
 
 class FleetController extends Controller
@@ -40,7 +40,7 @@ class FleetController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Fleet  $fleet
+     * @param  \App\Models\Fleet  $fleet
      * @return \Illuminate\Http\Response
      */
     public function show(Fleet $fleet)
@@ -51,7 +51,7 @@ class FleetController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Fleet  $fleet
+     * @param  \App\Models\Fleet  $fleet
      * @return \Illuminate\Http\Response
      */
     public function edit(Fleet $fleet)
@@ -63,7 +63,7 @@ class FleetController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Fleet  $fleet
+     * @param  \App\Models\Fleet  $fleet
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Fleet $fleet)
@@ -74,7 +74,7 @@ class FleetController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Fleet  $fleet
+     * @param  \App\Models\Fleet  $fleet
      * @return \Illuminate\Http\Response
      */
     public function destroy(Fleet $fleet)
