+++ app/Models/Activity.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Promotion extends Model
+class Activity extends Model
 {
     /**
      * The attributes that are mass assignable.
