+++ database/factories/ActivityFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Activity;
+use App\Models\Activity;
 use Faker\Generator as Faker;
 
 $factory->define(Activity::class, function (Faker $faker) {
