+++ database/factories/AgentFactory.php
@@ -2,12 +2,12 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Agent;
+use App\Models\Agent;
 use Faker\Generator as Faker;
 
 $factory->define(Agent::class, function (Faker $faker) {
     return [
-        'port_id' => factory(\App\Port::class),
+        'port_id' => factory(\App\Models\Port::class),
         'name' => $faker->name,
         'business_name' => $faker->word,
         'primary_contact_number' => $faker->word,
