+++ database/factories/CountryFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Country;
+use App\Models\Country;
 use Faker\Generator as Faker;
 
 $factory->define(Country::class, function (Faker $faker) {
