+++ database/factories/FleetFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
-use App\Fleet;
+use App\Models\Fleet;
 use Faker\Generator as Faker;
 
 $factory->define(Fleet::class, function (Faker $faker) {
