+++ database/factories/PortFactory.php
@@ -2,13 +2,13 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Port;
+use App\Models\Port;
 use Faker\Generator as Faker;
 
 $factory->define(Port::class, function (Faker $faker) {
     return [
         'name' => $faker->name,
-        'country_id' => factory(\App\Country::class),
+        'country_id' => factory(\App\Models\Country::class),
         'image' => $faker->word,
         'description' => $faker->text,
     ];
