+++ database/factories/PromotionFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Promotion;
+use App\Models\Promotion;
 use Faker\Generator as Faker;
 
 $factory->define(Promotion::class, function (Faker $faker) {
