+++ database/seeds/UsersTableSeeder.php
@@ -11,6 +11,6 @@ class UsersTableSeeder extends Seeder
      */
     public function run()
     {
-        factory(App\User::class, 10)->create();
+        factory(App\Models\User::class, 10)->create();
     }
 }
